Some code quality improvements
