+++ setup.py
@@ -28,7 +28,12 @@ setup(
         "Programming Language :: Python :: 3.12",
         "Operating System :: Microsoft :: Windows",
     ],
-    package_data={"hPyT": ["assets/*"]},
+    package_data={"hPyT": ["preview_app/*", "preview_app/assets/*"]},
+    entry_points={
+        "console_scripts": [
+            "hPyT-preview=hPyT.preview_app.cli:main",
+        ],
+    },
     keywords="Tkinter wxpython pyqt pyside GUI window controls decorations hide show titlebar border color animation",
     python_requires=">=3.6",
     project_urls={
